Poprawione nazewnictwo
