+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/invoicing/TaxCalculator.java
@@ -1,39 +1,6 @@
 package pl.com.bottega.ecommerce.sales.domain.invoicing;
 
-import java.math.BigDecimal;
+public interface TaxCalculator {
 
-import pl.com.bottega.ecommerce.sharedkernel.Money;
-
-public class TaxCalculator implements TaxInterface {
-
-    @Override
-    public Tax calculate(RequestItem item) {
-        Money net = item.getTotalCost();
-        BigDecimal ratio = null;
-        String desc = null;
-        switch (item.getProductData()
-                    .getType()) {
-            case DRUG:
-                ratio = BigDecimal.valueOf(0.05);
-                desc = "5% (D)";
-                break;
-            case FOOD:
-                ratio = BigDecimal.valueOf(0.07);
-                desc = "7% (F)";
-                break;
-            case STANDARD:
-                ratio = BigDecimal.valueOf(0.23);
-                desc = "23%";
-                break;
-
-            default:
-                throw new IllegalArgumentException(item.getProductData()
-                                                       .getType()
-                                                   + " not handled");
-        }
-        Money taxValue = net.multiplyBy(ratio);
-
-        return new Tax(taxValue, desc);
-    }
-
+    Tax calculate(RequestItem item);
 }
