--- src/main/java/pl/com/bottega/ecommerce/sales/domain/invoicing/TaxInterface.java
@@ -1,6 +1,0 @@
-package pl.com.bottega.ecommerce.sales.domain.invoicing;
-
-public interface TaxInterface {
-
-    Tax calculate(RequestItem item);
-}
